FIX-P2-1: why-ranked reconciles against an EXTERNAL recompute (R51)
The P2 work-order BLOCKER-1 headline (a term scored but omitted/zeroed in
the trace) does not reproduce on HEAD — FIX-H3 prints all six terms. But
the mandated regression shape (reconcile printed components against the
RETURNED score, fusion-free) went RED deterministically: digest-P0 printed
components recomputed to 0.8999999922069077 against a returned
0.8999999922069079. Root cause: Go permits FMA contraction inside the
scorer's additive expression, so on arm64 the returned score could differ
by 1-2 ulps from a plain IEEE-754 recompute of the printed trace — the
recompute every external auditor (python/bc) performs. The prior H3 test
missed it by reconciling against the printed total from the same stored
record, with a Go expression that fused identically.

Ruling R51 appended to RULINGS.md first (FIX-F5 process rule; numbered R51
because R50 is reserved for the parallel K1 work order): reconciliation is
defined against a plain IEEE-754 recompute in scorer term order; the scorer
must round per-term; every scored term must be printed.

Fix: rank.go score() wraps each product in an explicit float64 conversion
(a spec-guaranteed rounding barrier — no FMA), so the returned score is
bit-reproducible from the printed components in any language.

R46 sweep of the term-combination surface: rank.go score() (sole live
scoring site — Rank and RankUniformR) and rank/calibrate.go
WeightVector.score (§9.3 replay of stored components) — both gated; no
other combination site (grep over non-test code).

Regression (R51 mandatory shape, internal/daemon/fix_p2_1_test.go):
- TestP2WhyRankedReconcilesWithReturnedScore: parse why-ranked, recompute
  fusion-free from printed components+weights alone, assert exact equality
  with the RETURNED SearchOutput score (outside the explanation record —
  a scored-but-unprinted term always breaks it); all six term lines
  present; under P2, R/S/F/I/N all trace non-zero for the salient message.
- TestP2WhyRankedReconcilesDigestProfile: same over the rendered digest
  payload score= lines (the leg that was RED pre-fix). Both under P0+P2.
- fix_h3_test.go + retrieve_test.go recomputes aligned to the external-
  verifier semantics (the H3 test became timing-flaky in reverse once the
  scorer stopped fusing).

Live drill (rebuilt binary, throwaway mesh, real IPC): why-ranked traces
recomputed in PYTHON match the returned score bit-exactly under both
CAIRN_RANK_PROFILE=p2 (S/I/N non-zero) and P0.

make test + make vet green (reconciliation tests -count=5).
internal/log/ and reconcile/sync untouched.

+++ internal/daemon/fix_h3_test.go
@@ -51,8 +51,10 @@ func reconcile(t *testing.T, text string) {
 	}
 	v := func(name string) float64 { return rank.ParseDec(comp[name][0]) }
 	w := func(name string) float64 { return rank.ParseDec(comp[name][1]) }
-	sum := v("R")*w("R") + v("S")*w("S") + v("F")*w("F") +
-		v("P_eff")*w("P_eff") + v("I")*w("I") + v("N")*w("N")
+	// R51: plain IEEE recompute (explicit conversions forbid FMA), matching the
+	// external-verifier semantics the scorer now guarantees.
+	sum := float64(v("R")*w("R")) + float64(v("S")*w("S")) + float64(v("F")*w("F")) +
+		float64(v("P_eff")*w("P_eff")) + float64(v("I")*w("I")) + float64(v("N")*w("N"))
 	if got := rank.Dec(sum); got != total {
 		t.Fatalf("R47 VIOLATION: printed components recompute to %s, returned total %s\n%s", got, total, text)
 	}
